+++ app/src/main/java/com/sepproject/medicalmanagementapp/doctor/DoctorActivity.java
@@ -1,4 +1,4 @@
-package com.sepproject.medicalmanagementapp;
+package com.sepproject.medicalmanagementapp.doctor;
 
 import android.content.Intent;
 import android.support.v7.app.AppCompatActivity;
