+++ app/src/main/java/com/sepproject/medicalmanagementapp/doctor/DoctorActivityViewModel.java
@@ -1,4 +1,4 @@
-package com.sepproject.medicalmanagementapp;
+package com.sepproject.medicalmanagementapp.doctor;
 
 import android.arch.lifecycle.ViewModel;
