+++ app/src/main/java/com/sepproject/medicalmanagementapp/test/ViewTestActivity.java
@@ -1,4 +1,4 @@
-package com.sepproject.medicalmanagementapp.activity.test;
+package com.sepproject.medicalmanagementapp.test;
 
 public class ViewTestActivity {
 }
